Fraudapp/pyproject.toml
[project]
name = "fraudapp"
version = "0.1.0"
description = "\"Kafka-based Credit Card Fraud Detection\""
authors = [
    {name = "Shankara Narayana N G",email = "shankarnarayana92@gmail.com"}
]
requires-python = "^3.11"
dependencies = [
    "numpy (>=2.4.2,<3.0.0)",
    "pandas (>=3.0.1,<4.0.0)",
    "kafka-python (>=2.3.0,<3.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "pyyaml (>=6.0.3,<7.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false
